Move devolo home control base entity to separate module (#126042)
