+++ homeassistant/components/devolo_home_control/binary_sensor.py
@@ -14,7 +14,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import DevoloHomeControlConfigEntry
-from .devolo_device import DevoloDeviceEntity
+from .entity import DevoloDeviceEntity
 
 DEVICE_CLASS_MAPPING = {
     "Water alarm": BinarySensorDeviceClass.MOISTURE,
