+++ homeassistant/components/devolo_home_control/devolo_multi_level_switch.py
@@ -3,7 +3,7 @@
 from devolo_home_control_api.devices.zwave import Zwave
 from devolo_home_control_api.homecontrol import HomeControl
 
-from .devolo_device import DevoloDeviceEntity
+from .entity import DevoloDeviceEntity
 
 
 class DevoloMultiLevelSwitchDeviceEntity(DevoloDeviceEntity):
