+++ homeassistant/components/devolo_home_control/entity.py
