+++ homeassistant/components/devolo_home_control/sensor.py
@@ -15,7 +15,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import DevoloHomeControlConfigEntry
-from .devolo_device import DevoloDeviceEntity
+from .entity import DevoloDeviceEntity
 
 DEVICE_CLASS_MAPPING = {
     "battery": SensorDeviceClass.BATTERY,
